changes in orig repo

+++ ffcompiler/__init__.py
@@ -6,6 +6,8 @@ import yaml
 import subprocess
 
 """
+Configure VDC instance, or read config'd settings into class
+
 This will read a default yaml file and output to a instance-specific file, 
 to be read by processes as needed
 
@@ -32,11 +34,17 @@ class FFInstall():
         self.FFP = None
         #----
         self.FF_DIR = os.path.join(
-            os.path.dirname(os.path.abspath(__file__)), 
+            os.path.dirname(os.path.dirname(
+                os.path.abspath(__file__)
+                )
+            ), 
             'ENCODE_LIB'
             )
         self.ff_yaml = os.path.join(
-            os.path.dirname(os.path.abspath(__file__)), 
+            os.path.dirname(os.path.dirname(
+                os.path.abspath(__file__)
+                )
+            ), 
             'ffmpeg_binary.yaml'
             )
 
@@ -44,16 +52,18 @@ class FFInstall():
     def run(self):
         if not os.path.exists(self.FF_DIR):
             os.mkdir(self.FF_DIR)
-        
         self._PREP()
-        self._EXEC("""cd %s/ffmpeg-3.1.1
-            ./configure \
+        os.chdir(os.path.join(self.FF_DIR, 'ffmpeg-3.1.1'))
+
+        config_command = './configure \
                 --enable-gpl --enable-nonfree \
                 --disable-ffserver \
                 --disable-shared --enable-static \
-                --enable-libx264 --enable-libmp3lame --enable-pthreads
-            make
-        """ % (self.FF_DIR))
+                --enable-libx264 --enable-libmp3lame --enable-pthreads'
+
+        self._EXEC(command=config_command)
+        self._EXEC(command='make')
+
         self._CONNECT()
         self._GLOBALIZE()
 
@@ -72,28 +82,49 @@ class FFInstall():
             return False
 
 
+    def _EXEC(self, command):
+        # print "[cmd] %s" % command
+        process = subprocess.Popen(
+            command, 
+            stdout=subprocess.PIPE, 
+            stderr=subprocess.STDOUT, 
+            shell=True, 
+            universal_newlines=True
+            )
+        while True:
+            line = process.stdout.readline().strip()
+            if line == '' and process.poll() is not None:
+                break
+            sys.stdout.write('\r')
+            sys.stdout.write("%s" % (line.strip()))
+            sys.stdout.flush()
+        sys.stdout.write('')
+
+
     def _PREP(self):
+        if not os.path.exists(self.FF_DIR):
+            os.mkdir(self.FF_DIR)
         pkg_name = os.path.basename(urllib2.urlparse.urlparse(self.pkg_url).path)
         pkg_fpath = os.path.join(self.FF_DIR, pkg_name)
         tmp_fpath = os.path.join(self.FF_DIR, pkg_name + ".downloading")
+
+        if not os.path.exists(tmp_fpath):
+            return None
 
         with open(tmp_fpath, "wb") as f:
             print "downloading %s from %s" % (pkg_name, self.pkg_url)
             f.write(urllib2.urlopen(self.pkg_url).read())
             os.rename(tmp_fpath, pkg_fpath)
-
+            # print os.path.abspath(pkg_fpath)
         if pkg_name.endswith(".tar.bz2"):
             print "extracting %s" % pkg_name
-            self._EXEC("cd '%s' ; tar xjf %s" % (self.FF_DIR, pkg_name))
+            os.chdir(self.FF_DIR)
+            os.system('tar xjf %s' % (pkg_name))
 
         if pkg_name.endswith(".tar.gz"):
             print "extracting %s" % pkg_name
-            self._EXEC("cd '%s' ; tar xzf %s" % (self.FF_DIR, pkg_name))
-
-
-    def _EXEC(self, command):
-        print "[cmd] %s" % command
-        os.system(command)
+            os.chdir(self.FF_DIR)
+            os.system('tar xzf %s' % (pkg_name))
 
 
     def _CONNECT(self):
@@ -104,14 +135,17 @@ class FFInstall():
             if os.path.isfile(os.path.join(self.FF_DIR, f)):
                 os.remove(os.path.join(self.FF_DIR, f))
             if os.path.isdir(os.path.join(self.FF_DIR, f)):
-                self.FFM = os.path.join(self.FF_DIR, f, 'ffmpeg')
-                self.FFP = os.path.join(self.FF_DIR, f, 'ffprobe')
+                if 'ffmpeg' in f:
+                    self.FFM = os.path.join(self.FF_DIR, f, 'ffmpeg')
+                    self.FFP = os.path.join(self.FF_DIR, f, 'ffprobe')
+
 
     def _GLOBALIZE(self):
         ffdict = {
             'ffmpeg' : self.FFM,
             'ffprobe' : self.FFP
             }
+
         with open(self.ff_yaml, 'w') as outfile:
             outfile.write(
                 yaml.dump(
@@ -122,32 +156,13 @@ class FFInstall():
 
 
 
-def main():
-    # TODO: move to yaml, change to class activate
-    install_dict = {
-        "http://www.tortall.net/projects/yasm/releases/yasm-1.2.0.tar.gz": 'yasm-1.2.0',
-        "ftp://ftp.videolan.org/pub/x264/snapshots/x264-snapshot-20120302-2245-stable.tar.bz2" : 'x264-snapshot-20120302-2245-stable',
-        "http://downloads.sourceforge.net/lame/lame-3.99.5.tar.gz" : 'lame-3.99.5',
-        }
-
-    for key, entry in install_dict.iteritems():
-        F1 = FFInstall(
-            pkg_url=key
-            )
-        if F1.check() False:
-            F1._PREP()
-            F1._EXEC("""cd %s/%s
-                ./configure
-            """ % (F1.FF_DIR, entry))
-
-    F4 = FFInstall(
+if __name__ == '__main__':
+    FF = FFInstall(
         pkg_url="http://ffmpeg.org/releases/ffmpeg-3.1.1.tar.bz2"
         )
+    FF._PREP()
+    FF._CONNECT()
+    FF._GLOBALIZE()
+    print FF.FFM
 
-    if F4.check() is False:
-        F4.run()
-
-
-if __name__ == '__main__':
-    sys.exit(main())
